move Welcome to pages/
