+++ src/App.js
@@ -2,7 +2,7 @@ import React from "react";
 import { render } from "react-dom";
 import { Router } from "@reach/router";
 
-import Welcome from "./Welcome";
+import Welcome from "./pages/Welcome";
 import Map from "./pages/map/Map";
 
 render(
